@@ -15,7 +15,7 @@ publish_to: none
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   # The following adds the Cupertino Icons font to your application.
@@ -25,8 +25,9 @@ dependencies:
     sdk: flutter
 
 dev_dependencies:
-  e2e: ^0.4.0
   flutter_test:
+    sdk: flutter
+  integration_test:
     sdk: flutter
 
 # For information on the generic Dart part of this file, see the
