@@ -2,12 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:e2e/e2e.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:integration_test/integration_test.dart';
 import 'package:test_objc/dartle.dart';
 
 void main() {
-  E2EWidgetsFlutterBinding.ensureInitialized();
+  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
   testWidgets('simple call', (WidgetTester tester) async {
     final MessageSearchRequest request = MessageSearchRequest()
       ..query = 'Aaron';
@@ -31,7 +31,7 @@ void main() {
     final MessageSearchRequest request = MessageSearchRequest()
       ..query = 'error';
     final MessageApi api = MessageApi();
-    MessageSearchReply reply;
+    MessageSearchReply? reply;
     expect(() async {
       reply = await api.search(request);
     }, throwsException);
